+++ package-dev/Runtime/SentryInitialization.cs
@@ -1,23 +1,24 @@
 #if !UNITY_EDITOR
-#if UNITY_IOS
-#define SENTRY_NATIVE_IOS
+#if UNITY_IOS || (UNITY_STANDALONE_OSX && ENABLE_IL2CPP)
+#define SENTRY_NATIVE_COCOA
 #elif UNITY_ANDROID
 #define SENTRY_NATIVE_ANDROID
-#elif UNITY_STANDALONE_WIN && ENABLE_IL2CPP
-#define SENTRY_NATIVE_WINDOWS
+#elif ENABLE_IL2CPP && (UNITY_STANDALONE_WIN || UNITY_STANDALONE_LINUX)
+#define SENTRY_NATIVE
 #elif UNITY_WEBGL
 #define SENTRY_WEBGL
 #endif
 #endif
 
+using System;
 using UnityEngine;
 using UnityEngine.Scripting;
 
-#if SENTRY_NATIVE_IOS
+#if SENTRY_NATIVE_COCOA
 using Sentry.Unity.iOS;
-#elif UNITY_ANDROID
+#elif SENTRY_NATIVE_ANDROID
 using Sentry.Unity.Android;
-#elif SENTRY_NATIVE_WINDOWS
+#elif SENTRY_NATIVE
 using Sentry.Unity.Native;
 #elif SENTRY_WEBGL
 using Sentry.Unity.WebGL;
@@ -37,17 +38,36 @@ namespace Sentry.Unity
             {
                 var sentryUnityInfo = new SentryUnityInfo();
 
-#if SENTRY_NATIVE_IOS
-                SentryNativeIos.Configure(options);
+                Exception nativeInitException = null;
+
+                try
+                {
+#if SENTRY_NATIVE_COCOA
+                    SentryNativeCocoa.Configure(options);
 #elif SENTRY_NATIVE_ANDROID
-                SentryNativeAndroid.Configure(options, sentryUnityInfo);
-#elif SENTRY_NATIVE_WINDOWS
-                SentryNative.Configure(options);
+                    SentryNativeAndroid.Configure(options, sentryUnityInfo);
+#elif SENTRY_NATIVE
+                    SentryNative.Configure(options);
 #elif SENTRY_WEBGL
-                SentryWebGL.Configure(options);
+                    SentryWebGL.Configure(options);
 #endif
+                }
+                catch (DllNotFoundException e)
+                {
+                    nativeInitException = new Exception(
+                        "Sentry native-error capture configuration failed to load a native library. This usually " +
+                        "means the library is missing from the application bundle or the installation directory.", e);
+                }
+                catch (Exception e)
+                {
+                    nativeInitException = new Exception("Sentry native error capture configuration failed.", e);
+                }
 
                 SentryUnity.Init(options);
+                if (nativeInitException != null)
+                {
+                    SentrySdk.CaptureException(nativeInitException);
+                }
             }
         }
     }