+++ src/Sentry.Unity.Editor/Android/DebugSymbolUpload.cs
@@ -1,90 +1,185 @@
+using System;
 using System.IO;
+using System.Text.RegularExpressions;
+using Sentry.Extensibility;
 using Sentry.Unity.Integrations;
-using UnityEngine;
+using UnityEditor;
 
 namespace Sentry.Unity.Editor.Android
 {
-    internal static class DebugSymbolUpload
+    internal class DebugSymbolUpload
     {
-        internal static readonly string GradleExportedSymbolsPath = Path.Combine("unityLibrary", "src", "main", "symbols");
+        private readonly IDiagnosticLogger _logger;
 
-        public static string GetSymbolsPath(string unityProjectPath, string gradleProjectPath, bool isProjectExporting)
-        {
-            var symbolsPath = Path.Combine(unityProjectPath, "Temp", "StagingArea", "symbols");
-            if (isProjectExporting)
-            {
-                // We can no longer trust the Unity temp directory to exist so we copy the symbols to the export directory.
-                var copiedSymbolsPath = Path.Combine(gradleProjectPath, GradleExportedSymbolsPath);
-                CopySymbols(symbolsPath, copiedSymbolsPath);
+        internal const string RelativeBuildOutputPathOld = "Temp/StagingArea/symbols";
+        internal const string RelativeGradlePathOld = "Temp/gradleOut";
+        internal const string RelativeBuildOutputPathNew = "Library/Bee/artifacts/Android";
+        internal const string RelativeAndroidPathNew = "Library/Bee/Android";
 
-                return copiedSymbolsPath;
-            }
+        private readonly string _unityProjectPath;
+        private readonly string _gradleProjectPath;
 
-            return symbolsPath;
-        }
+        internal string[] _symbolUploadPaths;
 
-        public static void AppendUploadToGradleFile(string sentryCliPath, string gradleProjectPath, string symbolsDirectoryPath, IApplication? application = null)
-        {
-            application ??= ApplicationAdapter.Instance;
-
-            // Gradle doesn't support backslahes on path (Windows) so converting to forward slashes
-            if (application.Platform == RuntimePlatform.WindowsEditor)
-            {
-                sentryCliPath = sentryCliPath.Replace(@"\", "/");
-                symbolsDirectoryPath = symbolsDirectoryPath.Replace(@"\", "/");
-            }
-
-            if (!File.Exists(sentryCliPath))
-            {
-                throw new FileNotFoundException("Failed to find sentry-cli", sentryCliPath);
-            }
-
-            var gradleFilePath = Path.Combine(gradleProjectPath, "build.gradle");
-            if (!File.Exists(gradleFilePath))
-            {
-                throw new FileNotFoundException("Failed to find 'build.gradle'", gradleProjectPath);
-            }
-
-            if (!Directory.Exists(symbolsDirectoryPath))
-            {
-                throw new DirectoryNotFoundException($"Failed to find the symbols directory at {symbolsDirectoryPath}");
-            }
-
-            // There are two sets of symbols:
-            // 1. The one specific to gradle (i.e. launcher)
-            // 2. The ones Unity also provides via "Create Symbols.zip" that can be found in Temp/StagingArea/symbols/
-            // We either get the path to the temp directory or if the project gets exported we copy the symbols to the output directory
-            using var streamWriter = File.AppendText(gradleFilePath);
-            streamWriter.Write($@"
+        private string _symbolUploadTask = @"
 // Credentials and project settings information are stored in the sentry.properties file
 gradle.taskGraph.whenReady {{
     gradle.taskGraph.allTasks[-1].doLast {{
         println 'Uploading symbols to Sentry'
         exec {{
             environment ""SENTRY_PROPERTIES"", ""./sentry.properties""
-            executable ""{sentryCliPath}""
-            args = [""upload-dif"", ""./unityLibrary/src/main/jniLibs/""]
-        }}
-        exec {{
-            environment ""SENTRY_PROPERTIES"", ""./sentry.properties""
-            executable ""{sentryCliPath}""
-            args = [""upload-dif"", ""{symbolsDirectoryPath}""]
+            executable ""{0}""
+            args = [""upload-dif"", {1}]
         }}
     }}
-}}");
-        }
+}}";
 
-        internal static void CopySymbols(string sourcePath, string targetPath)
+        public DebugSymbolUpload(IDiagnosticLogger logger,
+            string unityProjectPath,
+            string gradleProjectPath,
+            bool isExporting = false,
+            IApplication? application = null)
         {
-            foreach (string dirPath in Directory.GetDirectories(sourcePath, "*", SearchOption.AllDirectories))
+            _logger = logger;
+
+            _unityProjectPath = unityProjectPath;
+            _gradleProjectPath = gradleProjectPath;
+
+            _symbolUploadPaths = GetSymbolUploadPaths(isExporting, application);
+        }
+
+        public void AppendUploadToGradleFile(string sentryCliPath)
+        {
+            var gradleFilePath = Path.Combine(_gradleProjectPath, "build.gradle");
+            if (!File.Exists(gradleFilePath))
             {
-                Directory.CreateDirectory(dirPath.Replace(sourcePath, targetPath));
+                throw new FileNotFoundException("Failed to find 'build.gradle'.", _gradleProjectPath);
             }
 
-            foreach (string newPath in Directory.GetFiles(sourcePath, "*", SearchOption.AllDirectories))
+            if (File.ReadAllText(gradleFilePath).Contains("sentry.properties"))
             {
-                File.Copy(newPath, newPath.Replace(sourcePath, targetPath), true);
+                _logger.LogDebug("Symbol upload has already been added in a previous build.");
+                return;
+            }
+
+            _logger.LogInfo("Appending debug symbols upload task to gradle file.");
+
+            sentryCliPath = ConvertSlashes(sentryCliPath);
+            if (!File.Exists(sentryCliPath))
+            {
+                throw new FileNotFoundException("Failed to find sentry-cli", sentryCliPath);
+            }
+
+            var symbolPathArgument = string.Empty;
+            foreach (var symbolUploadPath in _symbolUploadPaths)
+            {
+                if (Directory.Exists(symbolUploadPath))
+                {
+                    symbolPathArgument += $"\"{ConvertSlashes(symbolUploadPath)}\",";
+                }
+                else
+                {
+                    throw new DirectoryNotFoundException($"Failed to find the symbols directory at {symbolUploadPath}");
+                }
+            }
+
+            using var streamWriter = File.AppendText(gradleFilePath);
+            streamWriter.Write(_symbolUploadTask, sentryCliPath, symbolPathArgument);
+        }
+
+        public void RemoveUploadFromGradleFile()
+        {
+            var gradleFilePath = Path.Combine(_gradleProjectPath, "build.gradle");
+            if (!File.Exists(gradleFilePath))
+            {
+                throw new FileNotFoundException($"Failed to find 'build.gradle'.", _gradleProjectPath);
+            }
+
+            var gradleBuildFile = File.ReadAllText(gradleFilePath);
+            if (!gradleBuildFile.Contains("sentry.properties"))
+            {
+                _logger.LogDebug("Skipping removing the upload task. The task has not been added to the gradle project.");
+                return;
+            }
+
+            _logger.LogDebug("Removing the upload task from gradle project.");
+
+            // Replacing the paths with '.*' and escaping the task
+            var uploadTaskFilter = string.Format(_symbolUploadTask, ".*", ".*");
+            uploadTaskFilter = Regex.Replace(uploadTaskFilter, "\"", "\\\"");
+            uploadTaskFilter = Regex.Replace(uploadTaskFilter, @"\[", "\\[");
+
+            gradleBuildFile = Regex.Replace(gradleBuildFile, uploadTaskFilter, "");
+
+            using var streamWriter = File.CreateText(gradleFilePath);
+            streamWriter.Write(gradleBuildFile);
+        }
+
+        public void TryCopySymbolsToGradleProject(IApplication? application = null)
+        {
+            // The new building backend takes care of making the debug symbol files available within the exported project
+            if (IsNewBuildingBackend(application))
+            {
+                _logger.LogDebug("New building backend. Skipping copying of debug symbols.");
+                return;
+            }
+
+            _logger.LogInfo("Copying debug symbols to exported gradle project.");
+
+            var buildOutputPath = Path.Combine(_unityProjectPath, RelativeBuildOutputPathOld);
+            var targetRoot = Path.Combine(_gradleProjectPath, "symbols");
+
+            foreach (var sourcePath in Directory.GetFiles(buildOutputPath, "*.so", SearchOption.AllDirectories))
+            {
+                var targetPath = sourcePath.Replace(buildOutputPath, targetRoot);
+                _logger.LogDebug("Copying '{0}' to '{1}'", sourcePath, targetPath);
+
+                Directory.CreateDirectory(Path.Combine(Path.GetDirectoryName(targetPath)));
+                FileUtil.CopyFileOrDirectory(sourcePath, targetPath);
             }
         }
+
+        internal string[] GetSymbolUploadPaths(bool isExporting, IApplication? application = null)
+        {
+            if (isExporting)
+            {
+                _logger.LogInfo("Exporting the project. Root for symbols upload: {0}", _gradleProjectPath);
+                return new[] { _gradleProjectPath };
+            }
+
+            if (IsNewBuildingBackend(application))
+            {
+                _logger.LogInfo("Unity version 2021.2 or newer detected. Root for symbols upload: 'Library'.");
+                return new[]
+                {
+                    Path.Combine(_unityProjectPath, RelativeBuildOutputPathNew),
+                    Path.Combine(_unityProjectPath, RelativeAndroidPathNew)
+                };
+            }
+
+            _logger.LogInfo("Unity version 2021.1 or older detected. Root for symbols upload: 'Temp'.");
+            return new[]
+            {
+                Path.Combine(_unityProjectPath, RelativeBuildOutputPathOld),
+                Path.Combine(_unityProjectPath, RelativeGradlePathOld)
+            };
+        }
+
+        internal static bool IsNewBuildingBackend(IApplication? application = null)
+        {
+            application ??= ApplicationAdapter.Instance;
+
+            // Starting from 2021.2 Unity caches the build output inside 'Library' instead of 'Temp'
+            var version = new Version(application.UnityVersion.Substring(0, 6)); // year.version
+            if (version >= new Version("2021.2"))
+            {
+                return true;
+            }
+
+            return false;
+        }
+
+        // Gradle doesn't support backslashes on path (Windows) so converting to forward slashes
+        internal static string ConvertSlashes(string path) => path.Replace(@"\", "/");
     }
 }