+++ test/Sentry.Unity.Editor.Tests/ScriptableSentryUnityOptionsTests.cs
@@ -44,6 +44,8 @@ namespace Sentry.Unity.Editor.Tests
             StringAssert.Contains("IosNativeSupportEnabled", optionsAsString);
             StringAssert.Contains("AndroidNativeSupportEnabled", optionsAsString);
             StringAssert.Contains("WindowsNativeSupportEnabled", optionsAsString);
+            StringAssert.Contains("MacosNativeSupportEnabled", optionsAsString);
+            StringAssert.Contains("LinuxNativeSupportEnabled", optionsAsString);
             StringAssert.Contains("OptionsConfiguration", optionsAsString);
             StringAssert.Contains("Debug", optionsAsString);
             StringAssert.Contains("DebugOnlyInEditor", optionsAsString);